refactor: extract `getUploader` from `index.ts`
